Refactor code to use ES6 module syntax
