+++ package.json
@@ -2,11 +2,15 @@
   "name": "@masatomakino/canvas-particle-waypoint",
   "version": "0.6.0",
   "description": "Particle waypoint module for canvas 2d",
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "module": "./esm/index.js",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./esm/index.d.ts",
+      "default": "./esm/index.js"
+    }
+  },
+  "types": "./esm/index.d.ts",
   "files": [
-    "lib",
     "esm"
   ],
   "repository": {
@@ -35,11 +39,11 @@
     "start:dev": "npm run server & npm run watch:tsc & npm run watch:demo",
     "build": "npm run buildTS && npm run demo && npm run typedocs",
     "demo": "npx @masatomakino/gulptask-demo-page",
-    "buildTS": "tsc --project tsconfig.json & tsc --project tsconfig.esm.json",
+    "buildTS": "tsc",
     "typedocs": "npx typedoc --out ./docs/api src/index.ts",
     "server": "npx browser-sync ./docs/demo -w",
     "watch:demo": "npm run demo -- -W",
-    "watch:tsc": "tsc -W --incremental --project tsconfig.json & tsc -W --incremental --project tsconfig.esm.json"
+    "watch:tsc": "npx tsc -W --incremental"
   },
   "bugs": {
     "url": "https://github.com/MasatoMakino/canvas-particle-waypoint/issues"
@@ -47,6 +51,6 @@
   "homepage": "https://github.com/MasatoMakino/canvas-particle-waypoint#readme",
   "directories": {
     "doc": "docs",
-    "lib": "lib"
+    "lib": "esm"
   }
 }
