--- tsconfig.esm.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "target": "es6",
-    "module": "es2015",
-    "outDir": "./esm",
-    "declaration": false,
-    "declarationMap": false
-  }
-}
